+++ vision/detectors/__init__.py
@@ -1,5 +1,7 @@
 from .color import ColorDetector
+from .face import FaceDetector
 
 __all__ = [
     "ColorDetector",
+    "FaceDetector",
 ]